# openapi.yaml
openapi: 3.1.0
info:
  title: Redubber API
  description: '# Redubber API - AI-Powered Video Redubbing Platform


    The Redubber API provides comprehensive video redubbing capabilities powered by
    OpenAI''s

    Whisper, GPT-4, and TTS technologies. Transform videos into different languages
    while

    preserving the original speaker''s tone, emotion, and delivery style.


    ## Key Features


    - **Intelligent Voice Refinement**: AI-powered voice instruction generation for
    natural-sounding TTS

    - **Multi-Voice Testing**: Generate previews for all available voices simultaneously

    - **Smart Caching**: Avoid redundant TTS generations with intelligent caching

    - **Async Processing**: High-performance async TTS with 5x throughput improvement

    - **Project Management**: Organize videos, track progress, and manage redubbing
    workflows


    ## Workflow


    1. Create a project and upload videos

    2. Start redubbing task (transcribe, translate, generate audio)

    3. Use voice refinement to test different voices and instructions

    4. Select the best voice and apply to all segments

    5. Download the final redubbed video


    ## Rate Limits


    - OpenAI API limits apply to transcription, translation, and TTS operations

    - Recommended: 1-3 concurrent redubbing tasks per instance

    - TTS generation: Up to 100 concurrent requests (configurable)'
  contact:
    name: Redubber Development Team
...
    email: support@redubber.io
  license:
...
paths:
  /api/projects/:
    get:
      tags:
      - projects
      summary: List Projects
      description: "List all projects ordered by last update.\n\nArgs:\n    db: DatabaseManager\
        \ dependency.\n\nReturns:\n    List of all projects, most recently updated\
        \ first."
      operationId: list_projects_api_projects__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProjectResponse'
                type: array
                title: Response List Projects Api Projects  Get
    post:
      tags:
      - projects
      summary: Create Project
      description: "Create or open a project and trigger file scan.\n\nCreates a new\
        \ project entry in the database if it doesn't exist,\nor updates the timestamp\
        \ if it already exists. Automatically triggers\nan asynchronous file scan\
        \ to index video and subtitle files.\n\nArgs:\n    project: Project creation\
        \ request with directory path.\n    background_tasks: FastAPI background tasks\
        \ manager.\n    db: DatabaseManager dependency.\n    scanner: FileScanner\
        \ dependency.\n\nReturns:\n    Complete project information including generated\
        \ ID.\n\nRaises:\n    HTTPException: 400 if path is invalid or inaccessible.\n\
        \    HTTPException: 422 if path validation fails."
      operationId: create_project_api_projects__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}:
    get:
      tags:
      - projects
      summary: Get Project
      description: "Get detailed project information by ID.\n\nArgs:\n    project_id:\
        \ Unique project identifier.\n    db: DatabaseManager dependency.\n\nReturns:\n\
        \    Complete project information.\n\nRaises:\n    HTTPException: 404 if project\
        \ not found."
      operationId: get_project_api_projects__project_id__get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - projects
      summary: Delete Project
      description: "Delete project and all associated metadata.\n\nRemoves project\
        \ entry, video analysis records, and subtitle references.\nDoes not delete\
        \ actual video files on disk.\n\nArgs:\n    project_id: Project to delete.\n\
        \    db: DatabaseManager dependency.\n\nRaises:\n    HTTPException: 404 if\
        \ project not found."
      operationId: delete_project_api_projects__project_id__delete
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}/voice-settings:
    put:
      tags:
      - voice-refinement
      - voice-refinement
      summary: Save selected voice settings to project
      description: 'Update the project''s voice configuration with the selected voice
        and instructions.


        After testing multiple voices with previews, use this endpoint to:

        1. Save the selected voice (e.g., "nova", "shimmer")

        2. Save the finalized voice instructions

        3. Record which segment was used for testing (for audit trail)


        **Effects**:

        - Updates project''s default voice and instructions

        - All future TTS generations in this project will use these settings

        - Saves selection history in database for tracking

        - Returns updated project object with new voice settings


        **Workflow integration**:

        1. Analyze transcription → Generate voice instructions

        2. (Optional) Regenerate with feedback to refine instructions

        3. Generate previews for all voices

        4. User listens and selects best voice

        5. **Call this endpoint** to save selection

        6. All subsequent redubbing tasks use the saved settings


        **Use case**: Once voice selection is complete, this persists the choice

        so the entire project uses consistent voice settings across all videos.'
      operationId: save_voice_settings_api_projects__project_id__voice_settings_put
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoiceSaveRequest'
      responses:
        '200':
          description: Voice settings saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
              example:
                id: 42
                name: My Video Project
                path: /storage/projects/my_video_project
                created_at: '2026-07-01T10:30:00Z'
                updated_at: '2026-07-10T14:22:00Z'
                voice: nova
                voice_instructions: Speak with a warm, professional tone. Maintain
                  a moderate pace with clear enunciation.
        '404':
          description: Project not found
          content:
            application/json:
              example:
                detail: Project 42 not found
        '422':
          description: Invalid request parameters
          content:
            application/json:
              example:
                detail:
                - loc:
                  - body
                  - voice
                  msg: field required
                  type: value_error.missing
        '500':
          description: Failed to save voice settings (database error)
          content:
            application/json:
              example:
                detail: 'Failed to save voice settings: Database connection error'
  /api/projects/{project_id}/source-language:
    put:
      tags:
      - projects
      summary: Update Source Language
      description: "Update source language override for a project.\n\nSets or clears\
        \ the source language override. When set, all audio streams\nwill be treated\
        \ as this language during processing. This overrides the\nauto-detected language\
        \ from audio stream metadata.\n\nArgs:\n    project_id: Target project identifier.\n\
        \    language_update: New source language setting.\n    db: DatabaseManager\
        \ dependency.\n\nReturns:\n    Updated project information.\n\nRaises:\n \
        \   HTTPException: 404 if project not found."
      operationId: update_source_language_api_projects__project_id__source_language_put
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceLanguageUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProjectCreate:
      properties:
        path:
          type: string
          title: Path
          description: Absolute path to project directory
      type: object
      required:
      - path
      title: ProjectCreate
      description: Request schema for creating/opening a project.
    ProjectResponse:
      properties:
        id:
          type: integer
          title: Id
          description: Unique project identifier
        path:
          type: string
          title: Path
          description: Absolute path to project directory
        name:
          type: string
          title: Name
          description: Project display name
        created_at:
          type: string
          format: date-time
          title: Created At
          description: Project creation timestamp
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: Last update timestamp
        voice:
          type: string
          title: Voice
          description: TTS voice identifier (e.g., 'nova', 'alloy')
          default: ''
        voice_instructions:
          type: string
          title: Voice Instructions
          description: Custom instructions for TTS voice
          default: ''
        source_language_override:
          type: string
          title: Source Language Override
          description: Auto-detected or user-overridden source language (e.g., 'rus',
            'eng')
          default: ''
      type: object
      required:
      - id
      - path
      - name
      - created_at
      - updated_at
      title: ProjectResponse
      description: Response schema for project information.
...
    SourceLanguageUpdate:
      properties:
        source_language:
          type: string
          maxLength: 10
          title: Source Language
          description: Source language code (e.g., 'eng', 'rus', 'zho'). Empty string
            to clear override.
      type: object
      required:
      - source_language
      title: SourceLanguageUpdate
      description: Request schema for updating project source language.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    VoiceSaveRequest:
      properties:
        voice:
          type: string
          title: Voice
          description: Selected voice name (e.g., 'nova')
          examples:
          - nova
          - shimmer
          - alloy
          - echo
          - fable
          - onyx
        voice_instructions:
          type: string
          title: Voice Instructions
          description: Voice instructions to use
          examples:
          - Speak with a warm, professional tone. Maintain a moderate pace with clear
            enunciation.
          - Deliver with high energy and enthusiasm. Use a fast pace with dynamic
            intonation.
        segment_used:
          type: string
          title: Segment Used
          description: Segment ID that was used for voice testing
          examples:
          - project_42_segment_0
          - intro_video_segment_1
      type: object
      required:
      - voice
      - voice_instructions
      - segment_used
      title: VoiceSaveRequest
      description: Request schema for saving selected voice settings.
      examples:
      - segment_used: project_42_segment_0
        voice: nova
        voice_instructions: Speak with a warm, professional tone. Maintain a moderate
          pace with clear enunciation.
...